+++ test/query.cursor.test.js
@@ -209,9 +209,8 @@ describe('QueryCursor', function() {
     it('with pre-find hooks (gh-5096)', async function() {
       const schema = new Schema({ name: String });
       let called = 0;
-      schema.pre('find', function(next) {
+      schema.pre('find', function() {
         ++called;
-        next();
       });
 
       db.deleteModel(/Test/);
@@ -883,8 +882,8 @@ describe('QueryCursor', function() {
   it('throws if calling skipMiddlewareFunction() with non-empty array (gh-13411)', async function() {
     const schema = new mongoose.Schema({ name: String });
 
-    schema.pre('find', (next) => {
-      next(mongoose.skipMiddlewareFunction([{ name: 'bar' }]));
+    schema.pre('find', () => {
+      throw mongoose.skipMiddlewareFunction([{ name: 'bar' }]);
     });
 
     const Movie = db.model('Movie', schema);
@@ -905,6 +904,10 @@ describe('QueryCursor', function() {
 
   it('returns the underlying Node driver cursor with getDriverCursor()', async function() {
     const schema = new mongoose.Schema({ name: String });
+    // Add some middleware to ensure the cursor hasn't been created yet when `cursor()` is called.
+    schema.pre('find', async function() {
+      await new Promise(resolve => setTimeout(resolve, 10));
+    });
 
     const Movie = db.model('Movie', schema);
 
@@ -927,7 +930,7 @@ describe('QueryCursor', function() {
     const TestModel = db.model('Test', mongoose.Schema({ name: String }));
 
     const stream = await TestModel.find().cursor();
-    await once(stream, 'cursor');
+    assert.ok(stream.cursor);
     assert.ok(!stream.cursor.closed);
 
     stream.destroy();
@@ -939,7 +942,9 @@ describe('QueryCursor', function() {
 
   it('handles destroy() before cursor is created (gh-14966)', async function() {
     db.deleteModel(/Test/);
-    const TestModel = db.model('Test', mongoose.Schema({ name: String }));
+    const schema = mongoose.Schema({ name: String });
+    schema.pre('find', () => new Promise(resolve => setTimeout(resolve, 10)));
+    const TestModel = db.model('Test', schema);
 
     const stream = await TestModel.find().cursor();
     assert.ok(!stream.cursor);